Aula 8

+++ libs/boot.js
@@ -1,6 +1,8 @@
 module.exports = app => {
-    app.db.sequelize.sync({ force: true }) 
-    app.listen(app.get("port"), () => {
-        console.log('NTask API = porta ', app.get("port"));
-    });
+    if (process.env.NODE_ENV !== "test"){
+           app.db.sequelize.sync({ force: true }) 
+           app.listen(app.get("port"), () => {
+                console.log('NTask API = porta ', app.get("port"));
+            });
+    }
 };
